pyproject.toml
[project]
name = "bidding-docs-similarity"
version = "0.1.0"
description = "Detect similarity of bidding documents"
authors = [
    { name = "Developer", email = "developer@example.com" },
]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-multipart = "^0.0.6"
pdfplumber = "^0.10.2"
pymupdf = "^1.23.8"
python-docx = "^0.8.11"
sentence-transformers = "^2.2.2"
torch = "^2.1.0"
faiss-cpu = "^1.7.4"
numpy = "^1.24.4"
language-tool-python = "^2.7.1"
openpyxl = "^3.1.2"
paddlepaddle = "^2.5.2"
paddleocr = "^2.7.0.3"
opencv-python = "^4.8.1.78"
python-dotenv = "^1.0.0"
celery = "^5.3.4"
redis = "^5.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"